requirements.txt
cx-Freeze==6.11.1
packaging==21.3
patchelf==0.15.0.0
pika==1.3.0
psutil==5.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
